pyproject.toml
[project]
name = "adam"
version = "0.1.0"
description = "Assistant Digital pour Applications Multiples"
authors = [
    {name = "Brice Harper"}
]
requires-python = "3.13.5"
dependencies = [
    "django (>=6.0.4,<7.0.0)",
    "django-quill-editor (>=0.1.42,<0.2.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "requests (>=2.33.1,<3.0.0)",
    "pillow (>=12.2.0,<13.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
